Module calyx::frontend::ast

Abstract Syntax Tree for Calyx

Structs

BitNum

Custom bitwidth numbers

Cell

The Cell AST nodes.

ComponentDef

AST statement for defining components.

Group

Guard

Guards expr using the optional guard condition guard.

NamespaceDef

Corresponds to an individual Calyx file.

Proto

Prototype of the cell definition

Wire

Data for the -> structure statement.

Enums

Atom

Atomic operations used in guard conditions and RHS of the guarded assignments.

Control

Control AST nodes.

GuardComp

Possible comparison operators for guards.

GuardExpr

The AST for GuardExprs

NumType

Port

Statement that refers to a port on a subcomponent. This is distinct from a Portdef which defines a port.