Enum calyx::ir::Control

pub enum Control {
    Seq(Seq),
    Par(Par),
    If(If),
    While(While),
    Invoke(Invoke),
    Enable(Enable),
    Empty(Empty),
}

Control AST nodes.

Variants

Seq(Seq)

Tuple Fields

0: Seq

Represents sequential composition of control statements.

Par(Par)

Tuple Fields

0: Par

Represents parallel composition of control statements.

If(If)

Tuple Fields

0: If

Standard imperative if statement

While(While)

Tuple Fields

0: While

Standard imperative while statement

Invoke(Invoke)

Tuple Fields

0: Invoke

Invoke a sub-component with the given port assignments

Enable(Enable)

Tuple Fields

0: Enable

Runs the control for a list of subcomponents.

Empty(Empty)

Tuple Fields

0: Empty

Control statement that does nothing.

Implementations

impl Control

pub fn empty() -> Self

Convience constructor for empty.

pub fn seq(stmts: Vec<Control>) -> Self

Convience constructor for seq.

pub fn par(stmts: Vec<Control>) -> Self

Convience constructor for par.

pub fn enable(group: RRC<Group>) -> Self

Convience constructor for enable.

pub fn invoke(
comp: RRC<Cell>,
inputs: Vec<(Id, RRC<Port>)>,
outputs: Vec<(Id, RRC<Port>)>
) -> Self

Convience constructor for invoke.

pub fn if_(
port: RRC<Port>,
cond: Option<RRC<CombGroup>>,
tbranch: Box<Control>,
fbranch: Box<Control>
) -> Self

Convience constructor for if

pub fn while_(
port: RRC<Port>,
cond: Option<RRC<CombGroup>>,
body: Box<Control>
) -> Self

Convience constructor for while

impl Control

pub fn clone(con: &Control) -> Control

Associated clone method the control program. We don’t define this using the Clone trait because cloning control is not very common and clones should be explicit.