Enum calyx::ir::Guard

pub enum Guard {
    Or(Box<Guard>, Box<Guard>),
    And(Box<Guard>, Box<Guard>),
    Not(Box<Guard>),
    True,
    CompOp(PortComp, RRC<Port>, RRC<Port>),
    Port(RRC<Port>),
}

An assignment guard which has pointers to the various ports from which it reads.

Variants

Or(Box<Guard>, Box<Guard>)

Tuple Fields

0: Box<Guard>

1: Box<Guard>

Represents c1 || c2.

And(Box<Guard>, Box<Guard>)

Tuple Fields

0: Box<Guard>

1: Box<Guard>

Represents c1 && c2.

Not(Box<Guard>)

Tuple Fields

0: Box<Guard>

Represents !c1

True

The constant true

CompOp(PortComp, RRC<Port>, RRC<Port>)

Tuple Fields

0: PortComp

1: RRC<Port>

2: RRC<Port>

Comparison operator.

Port(RRC<Port>)

Tuple Fields

0: RRC<Port>

Uses the value on a port as the condition. Same as p1 == true

Implementations

impl Guard

Helper functions for the guard.

pub fn for_each<F>(&mut self, f: &mut F) where
F: FnMut(RRC<Port>) -> Option<Guard>,

Mutates a guard by calling f on every leaf in the guard tree and replacing the leaf with the guard that f returns.

pub fn all_ports(&self) -> Vec<RRC<Port>>

Returns all the ports used by this guard.

pub fn is_true(&self) -> bool

Returns true if this is a Guard::True.

pub fn update<F>(&mut self, upd: F) where
F: FnOnce(Guard) -> Guard,

Update the guard in place. Replaces this guard with upd(self). Uses std::mem::take for the in-place update.

pub fn op_str(&self) -> String

Return the string corresponding to the guard operation.

pub fn port(p: RRC<Port>) -> Self

pub fn and(self, rhs: Guard) -> Self

pub fn or(self, rhs: Guard) -> Self

pub fn eq(self, other: Guard) -> Self

pub fn neq(self, other: Guard) -> Self

pub fn le(self, other: Guard) -> Self

pub fn lt(self, other: Guard) -> Self

pub fn ge(self, other: Guard) -> Self

pub fn gt(self, other: Guard) -> Self