Module calyx::ir::rewriter

Structs

Rewriter

A structure to track rewrite maps for ports. Stores both cell rewrites and direct port rewrites. Attempts to apply port rewrites first before trying the cell rewrite.

Type Definitions

CellRewriteMap

Map to rewrite cell uses. Maps name of the old cell to the new ir::Cell instance.

CombGroupRewriteMap

Map name of old combination group to new combinational group

PortRewriteMap

Map to rewrite port uses. Maps the canonical name of an old port (generated using ir::Port::canonical) to the new ir::Port instance.