Struct calyx::ir::Cell

pub struct Cell {
    pub ports: SmallVec<[RRC<Port>; 10]>,
    pub prototype: CellType,
    pub attributes: Attributes,
    /* private fields */
}

Represents an instantiated cell.

Fields
ports: SmallVec<[RRC<Port>; 10]>

Ports on this cell

prototype: CellType

Underlying type for this cell

attributes: Attributes

Attributes for this group.

Implementations

impl Cell

pub fn new(name: Id, prototype: CellType) -> Self

Construct a cell

pub fn is_reference(&self) -> bool

Get a boolean describing whether the cell is external.

pub fn find<S>(&self, name: S) -> Option<RRC<Port>> where
S: Display + Clone + AsRef<str>,

Get a reference to the named port if it exists.

pub fn find_with_attr<S>(&self, attr: S) -> Option<RRC<Port>> where
S: AsRef<str>,

Get a reference to the first port that has the attribute attr.

pub fn find_all_with_attr<'a, S>(
&'a self,
attr: S
) -> impl Iterator<Item = RRC<Port>> + 'a where
S: AsRef<str> + 'a,

Return all ports that have the attribute attr.

pub fn get<S>(&self, name: S) -> RRC<Port> where
S: Display + Clone + AsRef<str>,

Get a reference to the named port and throw an error if it doesn’t exist.

pub fn is_component(&self) -> bool

Returns true iff this cell is an instance of a Calyx-defined component.

pub fn is_this(&self) -> bool

Returns true iff this cell is the signature of the current component

pub fn is_primitive<S>(&self, prim: Option<S>) -> bool where
S: AsRef<str>,

Returns true if this is an instance of a primitive. If the optional name is provided then only returns true if the primitive has the given name.

pub fn get_with_attr<S>(&self, attr: S) -> RRC<Port> where
S: AsRef<str>,

Get a reference to the first port with the attribute attr and throw an error if none exist.

pub fn type_name(&self) -> Option<&Id>

Returns the name of the component that is this cells type.

pub fn get_parameter<S>(&self, param: S) -> Option<u64> where
S: Display + Clone + AsRef<str>,

Get parameter binding from the prototype used to build this cell.

pub fn constant_name(val: u64, width: u64) -> Id

Return the canonical name for the cell generated to represent this (val, width) constant.

pub fn get_attribute<S>(&self, attr: S) -> Option<&u64> where
S: AsRef<str>,

Return the value associated with this attribute key.

pub fn add_attribute<S>(&mut self, attr: S, value: u64) where
S: Into<String>,

Add a new attribute to the group.

pub fn name(&self) -> &Id

Grants immutable access to the name of this cell.

pub fn ports(&self) -> &SmallVec<[RRC<Port>; 10]>

Returns a reference to all super::Port attached to this cells.

pub fn get_signature(&self) -> Vec<PortDef<u64>>