Trait calyx::ir::traversal::Visitor

pub trait Visitor {

    fn require_postorder() -> bool
    where
        Self: Sized,
    { ... }
    fn traverse_component(
        &mut self,
        comp: &mut Component,
        signatures: &LibrarySignatures,
        components: &[Component]
    ) -> CalyxResult<()>
    where
        Self: Sized,
    { ... }
    fn do_pass(&mut self, context: &mut Context) -> CalyxResult<()>
    where
        Self: Sized + ConstructVisitor,
    { ... }
    fn do_pass_default(context: &mut Context) -> CalyxResult<Self>
    where
        Self: ConstructVisitor + Sized,
    { ... }
    fn start(
        &mut self,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish(
        &mut self,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_seq(
        &mut self,
        _s: &mut Seq,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_seq(
        &mut self,
        _s: &mut Seq,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_par(
        &mut self,
        _s: &mut Par,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_par(
        &mut self,
        _s: &mut Par,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_if(
        &mut self,
        _s: &mut If,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_if(
        &mut self,
        _s: &mut If,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_while(
        &mut self,
        _s: &mut While,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_while(
        &mut self,
        _s: &mut While,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn enable(
        &mut self,
        _s: &mut Enable,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn invoke(
        &mut self,
        _s: &mut Invoke,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn empty(
        &mut self,
        _s: &mut Empty,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }

}

The visiting interface for a ir::Control program. Contains two kinds of functions:

start_: Called when visiting top-down.

finish_: Called when visiting bottow-up.

A pass will usually override one or more function and rely on the default visitors to automatically visit the children.

Provided Methods

fn require_postorder() -> boolwhere
Self: Sized,

Returns true if this pass requires a post-order traversal of the components. In a post-order traversal, if component B uses a component A, then A is guaranteed to be traversed before B.

fn traverse_component(
&mut self,
comp: &mut Component,
signatures: &LibrarySignatures,
components: &[Component]
) -> CalyxResult<()>where
Self: Sized,

Define the traversal over a component. Calls Visitor::start, visits each control node, and finally calls Visitor::finish.

fn do_pass(&mut self, context: &mut Context) -> CalyxResult<()>where
Self: Sized + ConstructVisitor,

Run the visitor on a given program ir::Context. The function mutably borrows the control program in each component and traverses it.

After visiting a component, it called ConstructVisitor::clear_data to reset the struct.

Panics

Panics if the pass attempts to use the control program mutably.

fn do_pass_default(context: &mut Context) -> CalyxResult<Self>where
Self: ConstructVisitor + Sized,

Build a Default implementation of this pass and call Visitor::do_pass using it.

fn start(
&mut self,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before the traversal begins.

fn finish(
&mut self,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after the traversal ends. This method is always invoked regardless of the Action returned from the children.

fn start_seq(
&mut self,
_s: &mut Seq,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::Seq node.

fn finish_seq(
&mut self,
_s: &mut Seq,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::Seq node.

fn start_par(
&mut self,
_s: &mut Par,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::Par node.

fn finish_par(
&mut self,
_s: &mut Par,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::Par node.

fn start_if(
&mut self,
_s: &mut If,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::If node.

fn finish_if(
&mut self,
_s: &mut If,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::If node.

fn start_while(
&mut self,
_s: &mut While,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::While node.

fn finish_while(
&mut self,
_s: &mut While,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::While node.

fn enable(
&mut self,
_s: &mut Enable,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed at an ir::Enable node.

fn invoke(
&mut self,
_s: &mut Invoke,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed at an ir::Invoke node.

fn empty(
&mut self,
_s: &mut Empty,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed at an ir::Empty node.