Struct calyx::passes::ComponentInterface

pub struct ComponentInterface;

Wires up the go and done holes for Calyx programs with a single enable to the component go and done ports.

For example:

component main(go: 1) -> (done: 1) {
    cells { .. }
    wires {
        group only_group { .. }
    }
    control { only_group; }
}

is transformed into:

component main(go: 1) -> (done: 1) {
    cells { .. }
    wires {
        group only_group { .. }
        only_group[go] = go;
        done = only_group[done];
    }
    control { only_group; }
}