Struct calyx::passes::RegisterUnsharing

pub struct RegisterUnsharing { /* fields omitted */ }

Unsharing registers reduces the amount of multiplexers used in the final design, trading them off for more memory.

A register use is said to be shared if it is used to store multiple, non-overlapping values in it. Unsharing, then, is the process of identifying such usages of registers and generating new registers to store non-overlapping values. For example, the following program:

let x = 1;
x = x + 2;
x = x + 3

Can be rewritten as:

let x = 1;
let y = x + 2;
let z = y + 3;

On the other hand, the following use of a register cannot be unshared:

let x = 0;
for i in 0..10 {
  x = x + 1;
}