Trait calyx::passes::ShareComponents

pub trait ShareComponents {
    fn lookup_node_conflicts(&self, node_name: &Id) -> Vec<&Id>;

    fn cell_filter(&self, cell: &Cell) -> bool;

    fn set_rewrites(&mut self, rewrites: HashMap<Id, RRC<Cell>>);

    fn get_rewrites(&self) -> &HashMap<Id, RRC<Cell>>;

    fn initialize(
        &mut self, 
        _component: &Component, 
        _library_signatures: &LibrarySignatures
    ) { ... }

    fn cell_equality(&self, cell0: &Cell, cell1: &Cell) -> bool { ... }

    fn custom_conflicts<F>(&self, _comp: &Component, _add_conflicts: F)
    where
        F: FnMut(Vec<Id>),
    { ... }
}

A trait for implementing passes that want to share components by building a conflict graph and performing graph coloring to minimize the number of used components.

You must implement the functions:

ShareComponents::lookup_node_conflicts

ShareComponents::cell_filter

ShareComponents::set_rewrites

ShareComponents::get_rewrites

Given these functions, the trait Visitor will automatically be implemented for your struct.

The algorithm that runs is:

instantiate conflict graph using all component cells that satisfy cell_filter

use ScheduleConflicts to find groups/invokes that run in parallel with each other

for each group/invoke, G that runs in parallel with another group/invoke H, add edges between each cell in the sets lookup_node_conflicts(G) and lookup_node_conflicts(H).

add conflicts between cells where for c0 != c1

call custom_conflicts to insert pass specific conflict edges

perform graph coloring using self.ordering to define the order of the greedy coloring

use coloring to rewrite group assignments, continuous assignments, and conditional ports.

Required methods

fn lookup_node_conflicts(&self, node_name: &Id) -> Vec<&Id>

Return a vector of conflicting cell names for a the group group_name. These are the names of the cells that conflict if their groups are run in parallel.

fn cell_filter(&self, cell: &Cell) -> bool

Given a cell and the library signatures, this function decides if this cell is relevant to the current sharing pass or not. This is used to filter out irrelevant cells.

fn set_rewrites(&mut self, rewrites: HashMap<Id, RRC<Cell>>)

Set the list of rewrites.

fn get_rewrites(&self) -> &HashMap<Id, RRC<Cell>>

Get the list of rewrites.

Provided methods

fn initialize(
&mut self,
_component: &Component,
_library_signatures: &LibrarySignatures
)

Initialize the structure. This function is called at the very beginning of the traversal before anything else.

fn cell_equality(&self, cell0: &Cell, cell1: &Cell) -> bool

The definition of cell equality. Cells will only be replaced with a cell that is equal to it according to this function. The default implementation is to compare the prototypes of the cell.

fn custom_conflicts<F>(&self, _comp: &Component, _add_conflicts: F) where
F: FnMut(Vec<Id>),

Called after the initial conflict graph is constructed. This function lets you add custom conflicts to the graph before graph coloring is performed.