Module calyx::utils

Shared utilities.

Modules

math

Structs

BoolIdx

Edge weight used for the graph nodes

FileIdx

Handle to a file in a PositionTable The index refers to the index in the PositionTable::files vector.

GPosIdx

A position index backed by a global PositionTable

GSym

GlobalPositionTable

The global position table

NameGenerator

Simple HashMap-based name generator that generates new names for each prefix.

PosIdx

Handle to a position in a PositionTable The index refers to the index in the PositionTable::indices vector.

PositionTable

Source position information for a Calyx program.

WeightGraph

Weight graph provides a wrapper over a Graph that allows adding edges using the NodeWeight type T (petgraph only allows adding edges using NodeIndex). Additionally, the edges are not allowed to have any weights.

Enums

OutputFile

Possible choices for output streams. Used by the -o option to the compiler.

Traits

WithPos

An IR node that may contain position information.

Type Definitions

Idx