Crate calyx

The Calyx Intermediate Language

Calyx is an intermediate language for transforming high-level programs into synthesizable hardware designs. Calyx’s key novelty is a split representation that captures the control-flow and the structural detail of a hardware design.

The control sub-language provides several high-level constructs: while (loops), if (conditionals), seq (sequencing), par (parallel execution), invoke (function calls). These contructs simplify the process of encoding the control-flow of a high-level program.

Calyx’s structural sub-language precisely capture details of the underlying hardware. Structural programs specify guarded assignments 1 using ports in structural components.

Take a look at the language tutorial for a complete overview.

Calyx’s guarded assignments are different from Bluespec’s rules. Rules can be dynamically aborted if there are conflicts at runtime and the Bluespec compiler generates scheduling logic to detect such cases. In contract, Calyx’s schedule is defined using the control program and requires no additional scheduling logic to detect aborts. ↩

Modules

analysis

Analysis for Calyx programs.

default_passes

Defines the default passes available to PassManager.

errors

Errors generated by the compiler.

frontend

Frontend parsing and AST representation.

ir

Internal representation for the Calyx compiler.

pass_manager

Define the PassManager structure that is used to construct and run pass passes.

passes

Passes for the Calyx compiler.

utils

Shared utilities.

Macros

build_assignments

Build guarded assignment statements and return a vector containing them.

guard

Convinience macro to get a port and turn it into an ir::Guard.

register_alias

Simple macro to register an alias with a pass manager.

structure

Add primitives and constants to the component and let-bind the references.