Enum calyx::ir::Control

pub enum Control {
    Seq(Seq),
    Par(Par),
    If(If),
    While(While),
    Invoke(Invoke),
    Enable(Enable),
    Empty(Empty),
}

Control AST nodes.

Variants

Seq(Seq)

Represents sequential composition of control statements.

Par(Par)

Represents parallel composition of control statements.

If(If)

Standard imperative if statement

While(While)

Standard imperative while statement

Invoke(Invoke)

Invoke a sub-component with the given port assignments

Enable(Enable)

Runs the control for a list of subcomponents.

Empty(Empty)

Control statement that does nothing.

Implementations

impl Control

pub fn empty() -> Self

Convience constructor for empty.

pub fn seq(stmts: Vec<Control>) -> Self

Convience constructor for seq.

pub fn par(stmts: Vec<Control>) -> Self

Convience constructor for par.

pub fn enable(group: RRC<Group>) -> Self

Convience constructor for enable.

pub fn invoke(
comp: RRC<Cell>,
inputs: Vec<(Id, RRC<Port>)>,
outputs: Vec<(Id, RRC<Port>)>
) -> Self

Convience constructor for invoke.

pub fn if_(
port: RRC<Port>,
cond: Option<RRC<CombGroup>>,
tbranch: Box<Control>,
fbranch: Box<Control>
) -> Self

Convience constructor for if

pub fn while_(
port: RRC<Port>,
cond: Option<RRC<CombGroup>>,
body: Box<Control>
) -> Self

Convience constructor for while

pub fn get_attribute<S>(&self, attr: S) -> Option<&u64> where
S: Display + AsRef<str>,

Returns the value of an attribute if present

pub fn has_attribute<S>(&self, attr: S) -> bool where
S: Display + AsRef<str>,

Returns true if the node has a specific attribute

impl Control

pub fn clone(con: &Control) -> Control

Associated clone method the control program. We don’t define this using the Clone trait because cloning control is not very common and clones should be explicit.