Struct calyx::ir::Builder

pub struct Builder<'a> {
    pub component: &'a mut Component,
    // some fields omitted
}

IR builder. Uses internal references to the component to construct and validate constructs when needed. By default, assumes that the cells are being added by a pass and marks them with the @generated attribute.

In order to disable this behavior, call [ir::Builder::not_generated()].

Fields
component: &'a mut Component

Component for which this builder is constructing.

Implementations

impl<'a> Builder<'a>

pub fn new(component: &'a mut Component, lib: &'a LibrarySignatures) -> Self

Instantiate a new builder using for a component.

pub fn validate(self) -> Self

Enable the validation flag on a builder.

pub fn not_generated(self) -> Self

Disable the generated flag on the builder

pub fn add_group<S>(&mut self, prefix: S) -> RRC<Group> where
S: Into<Id> + ToString + Clone,

Construct a new group and add it to the Component. The group is guaranteed to start with prefix. Returns a reference to the group.

pub fn add_comb_group<S>(&mut self, prefix: S) -> RRC<CombGroup> where
S: Into<Id> + ToString + Clone,

Construct a combinational group

pub fn add_constant(&mut self, val: u64, width: u64) -> RRC<Cell>

Return reference for a constant cell associated with the (val, width) pair, building and adding it to the component if needed.. If the constant does not exist, it is added to the Context.

pub fn add_primitive<Pre, Prim>(
&mut self,
prefix: Pre,
primitive: Prim,
param_values: &[u64]
) -> RRC<Cell> where
Pre: Into<Id> + ToString + Clone,
Prim: AsRef<str>,

Consturcts a primitive cell of type primitive. The name of the cell is guaranteed to start with prefix. Adds this cell to the underlying component and returns a reference to the Cell.

For example:

// Construct a std_reg.
builder.add_primitive("fsm", "std_reg", vec![32]);

pub fn add_component<Pre>(
&mut self,
prefix: Pre,
component: Pre,
sig: Vec<(Id, u64, Direction, Attributes)>
) -> RRC<Cell> where
Pre: Into<Id> + ToString + Clone,

Add a component instance to this component using its name and port signature.

pub fn build_assignment(
&self,
dst: RRC<Port>,
src: RRC<Port>,
guard: Guard
) -> Assignment

Construct an assignment.