Module calyx::ir::traversal

Visitor to traverse a control program.

Structs

CompTraversal

Define traversal order of components: pre-order, post-order, or none.

Enums

Action

Action performed at the end of visiting a control statement.

Order

The order in which the components are traversed.

Traits

ConstructVisitor

Trait defining method that can be used to construct a Visitor from an ir::Context. This is useful when a pass needs to construct information using the context before visiting the components.

Named

Trait that describes named things. Calling do_pass and do_pass_default. require this to be implemented.

Visitable

Describes types that can be visited by things implementing Visitor. This performs a recursive walk of the tree.

Visitor

The visiting interface for a ir::Control program. Contains two kinds of functions:

Type Definitions

VisResult

Result of performing a visit.