Trait calyx::ir::traversal::Named

pub trait Named {
    fn name() -> &'static str;

    fn description() -> &'static str;
}

Trait that describes named things. Calling do_pass and do_pass_default. require this to be implemented.

This has to be a separate trait from Visitor because these methods don’t recieve self which means that it is impossible to create dynamic trait objects.

Required methods

fn name() -> &'static str

The name of a pass. Is used for identifying passes.

fn description() -> &'static str

A short description of the pass.